+++ ui/src/admin/components/AdminTabs.js
@@ -88,11 +88,17 @@ const AdminTabs = ({
   return (
     <Tabs className="row">
       <TabList customClass="col-md-2 admin-tabs">
-        {tabs.map((t, i) => <Tab key={tabs[i].type}>{tabs[i].type}</Tab>)}
+        {tabs.map((t, i) =>
+          <Tab key={tabs[i].type}>
+            {tabs[i].type}
+          </Tab>
+        )}
       </TabList>
       <TabPanels customClass="col-md-10 admin-tabs--content">
         {tabs.map((t, i) =>
-          <TabPanel key={tabs[i].type}>{t.component}</TabPanel>
+          <TabPanel key={tabs[i].type}>
+            {t.component}
+          </TabPanel>
         )}
       </TabPanels>
     </Tabs>