+++ ui/src/dashboards/reducers/ui.js
@@ -235,12 +235,12 @@ export default function ui(state = initialState, action) {
 
       const dashboards = state.dashboards.map(
         dashboard =>
-          (dashboard.id === dashboardID
+          dashboard.id === dashboardID
             ? {
                 ...dashboard,
                 templates: dashboard.templates.map(
                   template =>
-                    (template.id === templateID
+                    template.id === templateID
                       ? {
                           ...template,
                           values: values.map((value, i) => ({
@@ -249,10 +249,10 @@ export default function ui(state = initialState, action) {
                             type: TEMPLATE_VARIABLE_TYPES[template.type],
                           })),
                         }
-                      : template)
+                      : template
                 ),
               }
-            : dashboard)
+            : dashboard
       )
 
       return {...state, dashboards}