+++ ui/src/kapacitor/components/KapacitorRules.js
@@ -44,9 +44,8 @@ const KapacitorRules = ({
     )
   }
 
-  const tableHeader = rules.length === 1
-    ? '1 Alert Rule'
-    : `${rules.length} Alert Rules`
+  const tableHeader =
+    rules.length === 1 ? '1 Alert Rule' : `${rules.length} Alert Rules`
   return (
     <PageContents
       source={source}
@@ -55,7 +54,9 @@ const KapacitorRules = ({
       onCloseTickscript={onCloseTickscript}
     >
       <div className="panel-heading u-flex u-ai-center u-jc-space-between">
-        <h2 className="panel-title">{tableHeader}</h2>
+        <h2 className="panel-title">
+          {tableHeader}
+        </h2>
         <Link
           to={`/sources/${source.id}/alert-rules/new`}
           className="btn btn-sm btn-primary"
@@ -74,7 +75,7 @@ const KapacitorRules = ({
   )
 }
 
-const PageContents = ({children, source, tickscript, onCloseTickscript}) => (
+const PageContents = ({children, source, tickscript, onCloseTickscript}) =>
   <div className="page">
     <div className="page-header">
       <div className="page-header__container">
@@ -104,7 +105,6 @@ const PageContents = ({children, source, tickscript, onCloseTickscript}) => (
         />
       : null}
   </div>
-)
 
 const {arrayOf, bool, func, node, shape, string} = PropTypes